ARRISXB6-3458 : brlan0 mac address changing on every boot of XB6
Reason for change: Assiging the available erouter2 i.e CM+3 Mac address to brlan0 so that it is persistent across reboot.
Test Procedure: Build and verify.
Risks: CM mac address not availabe from arm side right after flashing the image. But subsequent reboot it does get the Mac address.
This kind of behaviour is seen in general as well.
Signed-off-by: Sangeeta_Das@comcast.com

Change-Id: Iab95b2370eaa6fcba09880cb48373b4334819b87