CISCOXB3-2070: Incorrect boot_to_ETH_uptime and boot_to_WIFI_uptime values
Reason for change: In cases where wan comes later than LAN gw_lan_refresh is not called
as part of LAN initialization. So logging boot_to_ETH_uptime and boot_to_WIFI_uptime values
as part of WAN initialization if not done as part of LAN side initialization.
Test Procedure: Power on the box and let it boot completely and
Check whether the numbers in /rdklogs/logs/BootTime.log and ping to internet / wifi LED timers are matching
Risks: There shouldnt be any significant increase in the boot time.
Signed-off-by: Raghavendra Kammara <raghavendra_kammara@cable.comcast.com>

Change-Id: I1c6d26b2241d92f34b29d689ae97e9a08551de20