RDKB-12368: Fix typo on dns_sync script
Reason for change: It has been taking longer than usual for the ssl certs
to be installed causing scp to fail. Fixed an issue with the dns_sync.sh
script and increased the rety count

Test Procedure: Verified that the scp was successful after a number of
retries

Risks: None.

Change-Id: I5c7a00ad624d91a71ea92c5ad41bbccefee52a56
Signed-off-by: scott_angel@cable.comcast.com