ARRISXB6-5080 TCXB6-2404 ARRISXB6-5358 out of sync mesh vaps
Reason for change: Delete existing mesh vap configuration after each
Quantenna firmware boot. The Quantenna firmware boots up with a fixed
configuration for ath12 and ath13, so the stored configuration is no
longer valid. We choose to configure the VAPs in the firmware because of
ARRISXB6-4938.

Test procedure: Boot once, run:
- qwcfg_test set 6 ssid 'badssid'
- reboot
- observe qwcfg_test get 6 ssid is 'we.piranha.off' not 'badssid'

Risks: None.

Change-Id: Idb759d96ee38010d8d61fc69ab77aae2c9c49dce
Signed-off-by: Joe Bormolini <jbormolini@quantenna.com>