RDKB-14523: Clear mesh state on factory reset
Reason for change: On XB3 platforms, there are 2 syscfg databases. One
on the ATOM side as well as ARM. Factory reset only clears out the ARM
version so we ned to do something special to clear out the one on the
ATOM side.

Test Procedure: Verified that a PIN reset will remove the syscfg database
on the ATOM side.

Risks: Low.

Change-Id: I793673fed1880338d1a8ee44a1c3942243cdb442
Signed-off-by: scott_angel@cable.comcast.com