RDKB-15379: Disable the exposed zebra (vty) listener.
Reason for change:zebra (vty) listened on TCP port 2601.The listening port
was the CLI for Zebra, which should not need to be exposed on the LAN.
Inorder to disable the vty/console, the vty port is set to 0 from the CLI.
Test Procedure: Use netstat; Use tcpdump to capture the router adv
packets.
Risks: medium.
Signed-off-by: lsunny200 <livin_sunny2@comcast.com>

Change-Id: Ieadc701ebdd7895ffc6e6a34ffd2fbe43fcc2b58