RDKB-15530 : system() function call always returns -1 for success cases too.
Reason for change:
The daemon syseventd allows for registering event handlers. If they are executables, and if they use the system() API internally, those calls will always return -1. This is because system() expects to get the SIGCHLD event when the forked process finishes, but syseventd disables the SIGCHLD process(In syseventd_main.c, SIGCHLD signal is Ignored). This setting propagates to the event handlers, because they are child processes of syseventd or syseventd_fork_helper.
As a workaround on setting SIGCHLD back to SIG_DFL, system() function calls returns success on successful command execution.
Test Procedure:
1. WAN should work without any regression.
Risks: None
Signed-off-by: Lakshminarayanan Shenbagaraj Lakshminarayanan_Shenbagaraj@comcast.com

Change-Id: I7185f1fef9f78a608a74113a659e98880ec6f96f