RDKCOM-726: erouter0 was not getting IPv6 address intermittent when running...
Reason for change: erouter0 interface was not getting assigned IPv6 address [Intermittent].
Description: syseventd server is not completely up, so the client is not able to connect with that for setting default parameters.
Added the code to wait server to come up before proceeding to apply system default settings.
Test Procedure: 1) Build and test on XB6 and Puma7 platform.
2) Boot the box 20-30 times and check IPV6 on erouter0
Risks: None
Signed-off-by: Stephen Barrett <stephen.barrett@accenture.com>

Change-Id: Ie80ed210b885582a83a84cd3055773582e4a7f97