CISCOXB3-4078 dropbear is not running - Missing /var/run/dropbear.pid
Reason for change: The dropbear is a third party software component that is intended to create the PID file (if not PID (-P)) option is specified on creating the dropbear process. Intermittently it is not created. So -P option is included with the pid file path while starting the dropbear process.  If the PID file does not created then loged an error in Arm console log file.
Test Procedure: Start the dropbear. The /var/run/dropbear.pid file should be created.
Risks: None
Signed-off-by: Gokulakrishnan R <gokulakrishnan.r01@infosys.com>

Change-Id: I46a266031446e04e463cdf5bfdc071df01b4ce59