Logs not being rotated because of BOX_TYPE match
Reason for change: Logs in /nvram/log/firewall/ were not being rotated and
compressed because the BOX_TYPE variable was making the script chose the wrong
flock syntax.

Instead of making a workaround, flock can be used without the -c option,
making it work with a single call in any implementation.
Tested with flock from Busybox and GNU.

Test Procedure:
 1. Run log_handle.sh.
 2. Check that only fwlog.tar.bz2 remains in /nvram/log/firewall/.

Change-Id: I163b4021e2e80a8050e405b56e721fb32bdb0f67
Risks: Low