TCCBR-4920 TCCBR-4921 xfinity wifi going down after mode switch
Reason for change:
In CBR, erouter 0 goes down and comes up after sometime
(it takes around 65-80 seconds for it come back up)
during bridge - router and router - bridge transisitions,
if we query for "sysevent get wan-status" during this time,
we will get "stoppped". This sysevent on "Wan-status"
should be handled carefully, as this might down the ssid.
This was already did as a part of RDKB-23669, where we had
similar issue in XB6.
When the downed ssids are tried to be brought up by calling
set_ssids_enabled, which in turn calls get_ssids.
In get_ssids, the check is to add to localifs if and only if,
that ssid is currently enabled. as we have already have the
ssid in disabled state, they can't be brought back up.
So we are preventing the ssid from going on here

Make changes for Xfinity WiFi in similar lines to XB6
Test Procedure:
jira
Risks: High
Signed-off-by: Harikrishnan Raman <<Harikrishnan_Raman@comcast.com>>
Change-Id: I185bc93ef51e11580e1155888ec8f02e3b65ef59