RDKCOM-2389: RDKBDEV-628: fix misuse of "\0"
Reason for change:
"\0" is a pointer to an empty string (ie effectively the same as ""),
and is NOT the same as a nul charactor (ie a char with value 0).
Usage of "\0" is almost always a bug. If a pointer to an empty string
is needed then use "", if a nul character is needed then use 0 or
'\0'.
Test Procedure: Sanity.
Risks: None.
Signed-off-by: Andre McCurdy <armccurdy@gmail.com>

(cherry picked from commit 9ed11876d54910603a892a0e0b07b9064b920f03)
Change-Id: I2505cd32a83d3c88cc6ae4d3dabceb5ba14bc1e3