RDKCOM-2172: service_multinet: use fscanf() to parse psmcli output in nv_get...
Reason for change:
Using fscanf() is more efficient than the two step approach of using
fgets() to read the output into a temp buffer and then using sscanf()
to parse it. Using fscanf() also allows the 3 lines of output to be
parsed together (which isn't possible when using fgets() which stops
reading the input at each newline).
Apart from the improvement in efficiency, the main motivation for
this change is to improve the robustness of the parsing and avoid the
chance of crashing if the expected values are not found in psm (which
was seen in the previous version).
Test Procedure: Refer to JIRA
Risks: None
Signed-off-by: Andre McCurdy <armccurdy@gmail.com>

(cherry picked from commit 251421dd09874b7405b000e3a0047cfedcf6fafa)
Change-Id: I032ff7871c0d47c2359bfb4168dc99ebb1cf09a7