RDKB-32502: Observing syscfg_get failed in PAMlog.txt.0 logs
Reason for change: syscfg_get failed for hwHealthTest Enable gets printed
in PAM log file everytime there is a navigation change in UI. Since this
variable was not present in syscfg, this gets printed. Initialize a default
value (False) in the syscfg defaults file

Test Procedure: Navigate UI links and check the PAM log file. The mentioned
error should not be printed in it. Make sure that hwHealthTest variable was
not present before upgrading to build having this fix

Risks: Low

Signed-off-by: bharath_mallya@comcast.com
Change-Id: Ib3517c47732a778ff3f9d41d9c97c84ffa42e2a2