RDKBDEV-614: pmon: fix broken NULL pointer check in str_trim()
Reason for change:
The argument passed to str_trim() is the address of a pointer (the
address of a local variable within main() and so can never be NULL).

However, the pointer itself may be NULL (e.g. in the case of a
malformed config file with missing options, by the time str_trim() is
called for the third time, strsep() may have set the pointer to NULL)
and therefore needs to be checked.

Fixes crashes seen in pmon under long term stress testing.

Test Procedure: Sanity.
Risks: None.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: I0c3afb7089a98c8a2e4f1beb4d7a91422a3e5d8a