RDKBDEV-617: pmon.sh: fix variable name mixup in do_unregister()
Reason for change:
The variable being set was FEATURE but the second sysevent set
command uses $feature. Update the variable name so that the two
match.

Even with the bug everything probably worked OK since
pmon_feature_$ct was being cleared (and that's checked first to
determine if an entry is in use).

Test Procedure: Sanity.
Risks: None.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: I857a83b97ec25b8fe085daf2c2abb822091c2487