RDKBDEV-953: avoid unnecessary or repeated calls to strlen
Reason for change:
syscfg: avoid unnecessary or repeated calls to strlen(name)
The length of of the syscfg name parameter is retured by the initial
call to snprintf() to setup the name_p buffer. Storing that value
allows various calls to strlen(name) to be removed.

Test Procedure: Sanity.
Risks: None.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: Id779a583ddec70779bd6549ea22089b3ae81c0dc