RDKBDEV-1101: update rule for unwanted traffic to port 9000 from REJECT to DROP
Reason for change:

firewall: update rule for unwanted traffic to port 9000 from REJECT to DROP

Fix for cdrouter_firewall_100, cdrouter_firewall_101 and
cdrouter_firewall_110 which previously failed on Mv2+ with the error:

  FAIL: The following closed ports were found: 9000

Changed the target in the firewall rule for port 9000 from "REJECT"
to "DROP" so that packet is dropped and no response is sent back
during port scan.

Test Procedure: Sanity.
Risks: None.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: Ib5f588c51897bc1796ae661cfb071675cf1802aa