RDKB-37795 : Adding firewall rules to disable admin UI .
Reason for change: when remote management is disabled the admin UI should be disabled.
Test Procedure: As mentioned in the ticket
Risks: None

Signed-off-by: parvathy_sivaram@comcast.com

RDKB-37795 : Adding firewall rules to disable admin UI .

Reason for change: when remote management is enabled the admin UI is not able to access with http port.
Test Procedure: As mentioned in the ticket
Risks: None
Signed-off-by: gayathri_buddappagari@comcast.com

Change-Id: I50103a5340121819eb2177fbb23da2be874ac819